Some minor stuff
Changed rotation code to not...do whatever it was doing. Gotta test to see what joints positioning needs to be fixed for rotations.

MidiReader is only sending actions for the lower arms for an unknown reason....

Probably need to add more lines in the puppet creation to properly connect everything.... And still have to make the left leg... :100: